extract most of the code in index.ts into main.ts to make it easier to test
